Completely refactored naming scheme for Pedal related features
