+++ PedalCore/Pedal/Creation/PedalCreation.swift
@@ -1,0 +1,14 @@
+//
+//  PedalCreation.swift
+//  PedalCore
+//
+//  Created by Matheus Berger on 30/12/23.
+//
+
+import Foundation
+
+extension Pedal {
+    struct Creation {
+        // Namespace umbrella struct for PedalCreation related code
+    }
+}
