+++ PedalCore/Pedal/Creation/PedalCreationDelegate.swift
@@ -1,5 +1,5 @@
 //
-//  AddPedalDelegate.swift
+//  PedalCreationDelegate.swift
 //  PedalCore
 //
 //  Created by Lucas Migge on 02/10/23.
@@ -7,7 +7,7 @@
 
 import Foundation
 
-protocol CreatePedalDelegate: AnyObject {
+protocol PedalCreationDelegate: AnyObject {
     func finishedEditingPedal(_ pedal: Pedal.Model) throws
     
     func addNewPedal(_ pedal: Pedal.Model) throws
