+++ PedalCore/Pedal/List/PedalList.swift
@@ -1,0 +1,14 @@
+//
+//  PedalList.swift
+//  PedalCore
+//
+//  Created by Matheus Berger on 30/12/23.
+//
+
+import Foundation
+
+extension Pedal {
+    struct List {
+        // Namespace umbrella struct for PedalList related code
+    }
+}
