+++ PedalCore/Pedal/Models/Knob.swift
